refactor: build `db.createMigration` method with `buildCreateMigration`
