+++ packages/db-sqlite/src/index.ts
@@ -3,6 +3,7 @@ import type { DatabaseAdapterObj, Payload } from 'payload'
 
 import {
   beginTransaction,
+  buildCreateMigration,
   commitTransaction,
   count,
   countGlobalVersions,
@@ -37,6 +38,7 @@ import {
 } from '@payloadcms/drizzle'
 import { like } from 'drizzle-orm'
 import { createDatabaseAdapter, defaultBeginTransaction } from 'payload'
+import { fileURLToPath } from 'url'
 
 import type { Args, SQLiteAdapter } from './types.js'
 
@@ -44,12 +46,10 @@ import { connect } from './connect.js'
 import { countDistinct } from './countDistinct.js'
 import { convertPathToJSONTraversal } from './createJSONQuery/convertPathToJSONTraversal.js'
 import { createJSONQuery } from './createJSONQuery/index.js'
-import { createMigration } from './createMigration.js'
 import { defaultDrizzleSnapshot } from './defaultSnapshot.js'
 import { deleteWhere } from './deleteWhere.js'
 import { dropDatabase } from './dropDatabase.js'
 import { execute } from './execute.js'
-import { getMigrationTemplate } from './getMigrationTemplate.js'
 import { init } from './init.js'
 import { insert } from './insert.js'
 import { requireDrizzleKit } from './requireDrizzleKit.js'
@@ -57,6 +57,8 @@ import { requireDrizzleKit } from './requireDrizzleKit.js'
 export type { MigrateDownArgs, MigrateUpArgs } from './types.js'
 
 export { sql } from 'drizzle-orm'
+
+const filename = fileURLToPath(import.meta.filename)
 
 export function sqliteAdapter(args: Args): DatabaseAdapterObj<SQLiteAdapter> {
   const postgresIDType = args.idType || 'serial'
@@ -91,7 +93,6 @@ export function sqliteAdapter(args: Args): DatabaseAdapterObj<SQLiteAdapter> {
         json: true,
       },
       fieldConstraints: {},
-      getMigrationTemplate,
       idType: postgresIDType,
       initializing,
       localesSuffix: args.localesSuffix || '_locales',
@@ -122,7 +123,14 @@ export function sqliteAdapter(args: Args): DatabaseAdapterObj<SQLiteAdapter> {
       createGlobal,
       createGlobalVersion,
       createJSONQuery,
-      createMigration,
+      createMigration: buildCreateMigration({
+        filename,
+        sanitizeStatements({ sqlExecute, statements }) {
+          return statements
+            .map((statement) => `${sqlExecute}${statement?.replaceAll('`', '\\`')}\`)`)
+            .join('\n')
+        },
+      }),
       createVersion,
       defaultIDType: payloadIDType,
       deleteMany,
