+++ packages/db-sqlite/src/requireDrizzleKit.ts
@@ -1,15 +1,16 @@
 import type { RequireDrizzleKit } from '@payloadcms/drizzle/types'
 
-import { createRequire } from 'module'
-const require = createRequire(import.meta.url)
+import {
+  generateSQLiteDrizzleJson,
+  generateSQLiteMigration,
+  pushSQLiteSchema,
+} from 'drizzle-kit/api'
 
-/**
- * Dynamically requires the `drizzle-kit` package to access the `generateSQLiteDrizzleJson` and `pushSQLiteSchema` functions and exports them generically to call them from @payloadcms/drizzle.
- */
-export const requireDrizzleKit: RequireDrizzleKit = () => {
-  const {
-    generateSQLiteDrizzleJson: generateDrizzleJson,
-    pushSQLiteSchema: pushSchema,
-  } = require('drizzle-kit/api')
-  return { generateDrizzleJson, pushSchema }
-}
+export const requireDrizzleKit: RequireDrizzleKit = () => ({
+  // @ts-expect-error
+  generateDrizzleJson: generateSQLiteDrizzleJson,
+  // @ts-expect-error
+  generateMigration: generateSQLiteMigration,
+  // @ts-expect-error
+  pushSchema: pushSQLiteSchema,
+})
