+++ packages/drizzle/src/exports/postgres.ts
@@ -2,12 +2,10 @@ export { countDistinct } from '../postgres/countDistinct.js'
 export { createDatabase } from '../postgres/createDatabase.js'
 export { createExtensions } from '../postgres/createExtensions.js'
 export { createJSONQuery } from '../postgres/createJSONQuery/index.js'
-export { createMigration } from '../postgres/createMigration.js'
 export { defaultDrizzleSnapshot } from '../postgres/defaultSnapshot.js'
 export { deleteWhere } from '../postgres/deleteWhere.js'
 export { dropDatabase } from '../postgres/dropDatabase.js'
 export { execute } from '../postgres/execute.js'
-export { getMigrationTemplate } from '../postgres/getMigrationTemplate.js'
 export { init } from '../postgres/init.js'
 export { insert } from '../postgres/insert.js'
 export { migratePostgresV2toV3 } from '../postgres/predefinedMigrations/v2-v3/index.js'
