+++ packages/drizzle/src/postgres/predefinedMigrations/v2-v3/index.ts
@@ -3,7 +3,6 @@ import type { Payload, PayloadRequest } from 'payload'
 
 import { sql } from 'drizzle-orm'
 import fs from 'fs'
-import { createRequire } from 'module'
 import { buildVersionCollectionFields, buildVersionGlobalFields } from 'payload'
 import toSnakeCase from 'to-snake-case'
 
@@ -14,8 +13,6 @@ import type { PathsToQuery } from './types.js'
 import { groupUpSQLStatements } from './groupUpSQLStatements.js'
 import { migrateRelationships } from './migrateRelationships.js'
 import { traverseFields } from './traverseFields.js'
-
-const require = createRequire(import.meta.url)
 
 type Args = {
   debug?: boolean
@@ -43,8 +40,8 @@ export const migratePostgresV2toV3 = async ({ debug, payload, req }: Args) => {
   const dir = payload.db.migrationDir
 
   // get the drizzle migrateUpSQL from drizzle using the last schema
-  const { generateDrizzleJson, generateMigration, upPgSnapshot } = require('drizzle-kit/api')
-  const drizzleJsonAfter = generateDrizzleJson(adapter.schema)
+  const { generateDrizzleJson, generateMigration, upSnapshot } = adapter.requireDrizzleKit()
+  const drizzleJsonAfter = generateDrizzleJson(adapter.schema) as DrizzleSnapshotJSON
 
   // Get the previous migration snapshot
   const previousSnapshot = fs
@@ -63,8 +60,8 @@ export const migratePostgresV2toV3 = async ({ debug, payload, req }: Args) => {
     fs.readFileSync(`${dir}/${previousSnapshot}`, 'utf8'),
   ) as DrizzleSnapshotJSON
 
-  if (drizzleJsonBefore.version < drizzleJsonAfter.version) {
-    drizzleJsonBefore = upPgSnapshot(drizzleJsonBefore)
+  if (upSnapshot && drizzleJsonBefore.version < drizzleJsonAfter.version) {
+    drizzleJsonBefore = upSnapshot(drizzleJsonBefore)
   }
 
   const generatedSQL = await generateMigration(drizzleJsonBefore, drizzleJsonAfter)
