+++ packages/drizzle/src/postgres/requireDrizzleKit.ts
@@ -1,6 +1,11 @@
-import { createRequire } from 'module'
+import { generateDrizzleJson, generateMigration, pushSchema, upPgSnapshot } from 'drizzle-kit/api'
 
 import type { RequireDrizzleKit } from '../types.js'
-
-const require = createRequire(import.meta.url)
-export const requireDrizzleKit: RequireDrizzleKit = () => require('drizzle-kit/api')
+export const requireDrizzleKit: RequireDrizzleKit = () => ({
+  generateDrizzleJson,
+  generateMigration,
+  // @ts-expect-error
+  pushSchema,
+  // @ts-expect-error
+  upSnapshot: upPgSnapshot,
+})
