+++ packages/payload/src/database/types.ts
@@ -161,8 +161,6 @@ export type Connect = (args?: ConnectArgs) => Promise<void>
 export type Destroy = () => Promise<void>
 
 export type CreateMigration = (args: {
-  /** dirname of the package, required in drizzle */
-  dirname?: string
   file?: string
   forceAcceptWarning?: boolean
   migrationName?: string
